:bug: Renamed types => json_types for circular import fix
