+++ drawbot/bot.py
@@ -4,7 +4,7 @@ from typing import Optional
 
 from discord.ext import commands
 
-from .types import JsonData
+from .json_types import JsonData
 from .utils import json_wr
 
 
@@ -41,7 +41,7 @@ class Bot(commands.Bot):
         self.remove_command("help")
 
         for filename in os.listdir("drawbot/cogs"):
-            if filename.endswith(".py"):
+            if filename.endswith(".py") and filename != "__init__.py":
                 self.load_extension(f"drawbot.cogs.{filename[:-3]}")
 
     def run(self) -> None:
